Update intro.rst with a better producer/consumer
See gist for example test cases https://gist.github.com/wojtossfm/aede13446bb3b9e12942109f505058c1

The example code for the docs is prone to a race condition of the state of the tasks. If people replicate it they might wonder why they are losing messages (due to the cancel calls). The proposed change avoids the race condition.