Simplify connection termination on server shutdown.
Previously, there were two scenarios where a connection handler on the
server side would receive an unexpected connection termination:

1. When the remote endpoint closed the connection or it dropped for any
   reason: this was signalled with a ConnectionClosed exception when
   interacting with the connection, usually thrown by recv() since most
   connection handlers spend the majority of their time awaiting recv()

2. When the server shut down: this caused the connection handler to be
   cancelled, meaning that a CancelledError exception was injected at
   an arbitrary location.

Now, when the server shuts down, scenario 1 also applies. This removes
the need to prepare for CancelledError in connection handlers.

A good way to understand this change is to consider that "shutting down
the server" means "shutting down all connections".

Given that WebSocket connections are long-running, I find it sensible
to require connection handlers to detect when the connection they're
managing drops and to terminate.

I expect this to make it easier for users to write connection handlers
that behave correctly when the server shuts down, for two reasons:

- There's only one scenario to implement, which reduces the amount of
  exception handling required. In addition, this avoids the specially
  messy situation where the clean-up process for ConnectionClosed gets
  interrupted by CancelledError.
- Users are much more likely to check what happens when the connection
  drops than when the server shuts down. Making the latter behave like
  the former gives them proper behavior for free.

Another argument for this change is that cancellation wasn't properly
used there. Cancellation should only be used for terminating the
execution of a task that one started when one no longer cares about its
result. Since websockets runs a connection handler provided by the user,
it can't make the decision that whatever code comes next doesn't matter,
especially finalization code.

The counter-argument against this change is that shutdown will be much
slower for a server that reads events from an external source and sends
them over the WebSocket connection. Previously, such a server would
close immediately. Now, unless the connection handler checks for
termination explicitly, it won't notice that the server is closing
until it attempts to send the next event, which could take some time.

Fix #338 and #394.

Revert #337 and #392.