Let user override host/port from what is in uri when connecting
In some cases it is useful to connect using an IP address of a server, but you
might still need to keep proper hostname in the url for TLS or for vhosts
routing. For example if you are connecting to `wss://example.com` but via an ssh
tunnel you might want to set `host='127.0.0.1', port=50032` but then supply
`server_hostname='example.com'` for TLS validation.

With this change connection logic detects that alternative `host|port` is
supplied by the user and does not attempt to override it with values from the
uri. If secure connection is requested and host override is used then also set
`server_hostname` to the host name in the uri (unless `server_hostname` was
already supplied by user)