Change process_request to be a coroutine (again).
b64fee8e made it possible to use either a function or a coroutine. It
was part of the 7.0 release.

4f1a14c3 documented this possibility but wasn't released. However, users
may have read the "latest" docs and taken advantage of this.

For this reason, include proper deprecation warnings and preserve
backwards-compatibility (for the foreseeable future).

The deprecation warnings need to be in two locations to account for
passing a process_request argument and for overriding the
process_request method.

Issue #597 shows that `isinstance(..., Awaitable)` is more robust than
`asyncio.iscoroutinefunction(...)` because it also supports functions
returning awaitables.