Require Python ≥ 3.6.1.
There've been multiple regressions where websockets stops working with
Python 3.6 but works fine with Python 3.6.1 or higher. I don't have a
good way to detect such regressions. I don't think it's a good practice
to run anything but the latest 3.6.x (or 3.7.x, etc.) anyway. Instead of
fighting a useless and losing battle, I'm moving the minimum requirement
to 3.6.1.

Strictly speaking, this isn't a backwards incompatible change. The
incompatibility with Python 3.6 appeared in websockets 8.0, which was
the first release that included 94945fec.

Sure, I'm accepting the backwards incompatibility instead of fixing
it... Judge me if you'd like, or support websockets on Tidelift — if it
becomes profitable, I'll have an incentive to provide better support for
older Python versions.

Refs #655, #664, #667.