Advertise support for client_max_window_bits by default.
This doesn't change anything in the default setup (compression="deflate")
because it sets client_max_window_bits to 12 explicitly.

If a client is configured with extensions=[ClientPerMessageDeflateFactory()]
this change gives the server the option to limit client_max_window_bits.

Specifically, this makes such a client compatible with a websockets server
using the default setup (compression="deflate").