Add usage docs for start_serving and serve_forever.
The only example for Python I found was https://bugs.python.org/issue32662:

    async def main():
        srv = await asyncio.start_server(...)
        async with srv:
            await srv.serve_forever()

    asyncio.run(main())

It looks pretty bad to me: it starts the server thrice and stops it
twice, relying on idempotency to avoid issues. I dislike this style
of "pile it up to be sure that it works" so I'm showing a subset of
possibilities only.

Ref #1197.