Add API to customize subprotocol selection logic.
Also changed the default logic:

(1) to reject client connections that don't offer a subprotocol when the
server is configured with subprotocols.

This is the expected behavior for what I believe to be the default use
case: require one particular subprotocol.

This change of behavior isn't documented because I don't know anyone
embedding the Sans-I/O layer and supporting subprotocol selection at
this time.

(2) to rely only on the order of preference of the server.

Hey, trying to cater to the preferences of clients was nice, but the
behavior was so needlessly complex that documentation apologized...

This keeps things simple and still falls within the documented behavior.