don't use os.pipe with the I/O multiplexing selector on win32
On the Win32 platform, only sockets can be used with I/O
multiplexing (such as that performed by selectors.DefaultSelector);
the pipe cannot be added to the selector. However, on the win32
platform, simply closing the listener socket is enough to cause the
call to select to return -- the additional pipe is redundant.

On Mac OS X (and possibly other BSD derivatives), closing the listener
socket isn't enough. In the interest of maximum compatibility, we
simply disable the use of os.pipe on the Win32 platform.