Support closing while sending a fragmented message.
On one hand, it will close the connection with an unfinished fragmented
message, which is less than ideal.

On the other hand, RFC 6455 implies that it should be legal and it's
probably best to let users close the connection if they want to close
the connection (rather than force them to call fail() instead).