Add a respond method to server connections.
It's an alias of the reject method of the underlying server protocol.
It makes it easier to write process_request

It's called respond because the semantics are "consider the request
as an HTTP request and create an HTTP response".

There isn't a similar alias for accept because process_request should
just return and websockets will call accept.