Move broadcast() to the server module.
Rationale:

* It is only useful for servers. (Maybe there's a use case for client
  but I couldn't picture it.)
* It was already documented in the page covering the server module.
* Within this page, it was the only API from the connection or protocol
  module.

The implementation remains in the connection or protocol module because
moving it would require refactoring tests. I'd rather keep them simple.
(And I'm lazy.)

This change doesn't require a backwards compatibility shim because the
documentated location of the legacy implementation of broadcast was
websockets.broadcast, it's changing with the introduction of the new
asyncio API, and the changes are already documented.