Prevent close() from blocking when reading is paused.
Closing the transport normally is achieved with transport.write_eof().

Closing it in abnormal situations relied on transport.close(). However,
that didn't lead to connection_lost() when reading is paused.

Replacing it with transport.abort() ensures that buffers are dropped
(which is what we want in abnormal situations) and connection_lost()
called quickly.

Fix #1555 (for real!)